#30 feat: tablet 화면에서 날짜, 시간 선택 화면 모달로 분기 처리
